test: consolidate bias-gradient tests into per-operator test files
Move the D-RTRL-vs-BPTT bias-gradient checks out of the standalone
bias_gradient_test.py and into the test file for each operator they
exercise: dense (matmul + _mm_yw_to_w), conv (1D + 2D), and sparse.
Assertions and semantics are preserved; npt.* calls are converted to
the np.testing.* convention already used in the target files.
